establish app communication

+++ dashboard/src/main/kotlin/com/gigahack/dashboard/repository/ProductRepository.kt
@@ -3,5 +3,5 @@ package com.gigahack.dashboard.repository
 import com.gigahack.dashboard.model.Product
 import org.springframework.data.mongodb.repository.MongoRepository
 
-interface ProductRepository: MongoRepository<Product, Long> {
+interface ProductRepository: MongoRepository<Product, Int> {
 }
